Refactoring to make BlockKey a common component
