+++ cpp/tensorrt_llm/batch_manager/kvCacheManager.cpp
@@ -54,11 +54,6 @@ using BlocksPerWindow = std::map<SizeType32, std::tuple<SizeType32, SizeType32>>
 namespace
 {
 
-inline uint8_t getNthByte(SizeType32 hashPart, uint8_t byteIdx) noexcept
-{
-    return static_cast<uint8_t>((hashPart >> (24 - byteIdx * 8)) & 0xFF);
-}
-
 //! \brief Get all blocks in a sequence by traversing backwards from the last block.
 //! \param lastBlock is a BlockPtr to the last block in the sequence to start traversal from
 //! \return Vector of BlockPtr-s in sequence order
@@ -96,159 +91,6 @@ std::vector<BlockPtr> getAllSequenceBlocks(BlockPtr lastBlock)
 
 namespace tensorrt_llm::batch_manager::kv_cache_manager
 {
-std::vector<MmKey> generateBlockHashExtraKeys(
-    tensorrt_llm::batch_manager::LlmRequest const& llmRequest, SizeType32 startTokenIdx, SizeType32 endTokenIdx)
-{
-    auto const multimodalHashes = llmRequest.getMultimodalHashes();
-    auto const multimodalPositions = llmRequest.getMultimodalPositions();
-    auto const multimodalLengths = llmRequest.getMultimodalLengths();
-    auto const multimodalUuids = llmRequest.getMultimodalUuids();
-
-    if (!multimodalHashes || !multimodalPositions || !multimodalLengths || !(*multimodalHashes)
-        || (*multimodalHashes)->empty() || !(*multimodalPositions) || (*multimodalPositions)->empty()
-        || !(*multimodalLengths) || (*multimodalLengths)->empty())
-    {
-        return {};
-    }
-
-    if ((*multimodalHashes)->size() != (*multimodalPositions)->size()
-        || (*multimodalPositions)->size() != (*multimodalLengths)->size())
-    {
-        TLLM_LOG_WARNING("Multimodal data arrays have mismatched sizes");
-        return {};
-    }
-
-    std::vector<MmKey> extraKeys;
-    extraKeys.reserve((*multimodalPositions)->size());
-    std::array<uint8_t, 32> mmHashArray;
-
-    for (size_t i = 0; i < (*multimodalPositions)->size(); ++i)
-    {
-        auto const& startPos = (*(*multimodalPositions))[i];
-        auto const& length = (*(*multimodalLengths))[i];
-        auto const& mmHashVector = (*(*multimodalHashes))[i];
-
-        TLLM_CHECK_WITH_INFO(mmHashVector.size() == 8, "Multimodal hash vector has unexpected size: %zu (expected 8)",
-            mmHashVector.size());
-
-        // mmHashVector[j] comes from Python's int(hex_chunk, 16)
-        // where hex_chunk like "00010203" means 0x00 is MSB and 0x03 is LSB (big endian)
-        // Convert 8x 32-bit integers into a 32-byte array preserving Blake3 hash byte order
-        // Example: hashPart = 0x00010203 → mmHashArray[0:3] = [0x00, 0x01, 0x02, 0x03]
-        for (size_t j = 0; j < 8; ++j)
-        {
-            auto const& hashPart = mmHashVector[j];
-            for (uint8_t byteIdx = 0; byteIdx < 4; ++byteIdx)
-            {
-                mmHashArray[j * 4 + byteIdx] = getNthByte(hashPart, byteIdx);
-            }
-        }
-
-        // Check if this multimodal content overlaps with the current block
-        if (endTokenIdx > startPos && startTokenIdx < startPos + length)
-        {
-            uint64_t mmStartInBlock = (startPos >= startTokenIdx) ? 0 : static_cast<uint64_t>(startTokenIdx - startPos);
-
-            // Get UUID if available
-            std::optional<std::string> uuid = std::nullopt;
-            if (multimodalUuids && *multimodalUuids && i < (*multimodalUuids)->size())
-            {
-                uuid = (*(*multimodalUuids))[i];
-            }
-
-            extraKeys.emplace_back(mmHashArray, mmStartInBlock, std::move(uuid));
-        }
-    }
-
-    return extraKeys;
-}
-
-std::vector<BlockKey> buildBlockKeys(
-    std::list<VecUniqueTokens>& blockedUniqueTokens, tensorrt_llm::batch_manager::LlmRequest const& llmRequest)
-{
-    std::vector<BlockKey> blockKeys;
-
-    SizeType32 currentTokenIdx = 0;
-    for (auto& uniqueTokens : blockedUniqueTokens)
-    {
-        auto extraKeys = generateBlockHashExtraKeys(llmRequest, currentTokenIdx, currentTokenIdx + uniqueTokens.size());
-        currentTokenIdx += uniqueTokens.size();
-
-        blockKeys.emplace_back(llmRequest.getInputTokensExtraIds().has_value(), llmRequest.getLoraTaskId(),
-            std::move(uniqueTokens), std::move(extraKeys), llmRequest.getCacheSaltID());
-    }
-    return blockKeys;
-}
-
-bool BlockKey::operator==(BlockKey const& other) const noexcept
-{
-    return (usesExtraIds == other.usesExtraIds && loraTaskId == other.loraTaskId && uniqueTokens == other.uniqueTokens
-        && extraKeys == other.extraKeys && cacheSaltID == other.cacheSaltID);
-}
-
-BlockKey BlockKey::shorten(int newNumTokens) const
-{
-    TLLM_CHECK_WITH_INFO(
-        newNumTokens >= 0 && newNumTokens <= getNumTokens(), "newNumTokens must be >= 0 and <= getNumTokens()");
-    BlockKey result(*this);
-    result.uniqueTokens.resize(newNumTokens);
-    return result;
-}
-
-size_t BlockKeyHasher::hash(BlockKey const& blockKey, std::size_t parentHash) noexcept
-{
-    // Hashing algorithm adapted from StackOverflow:
-    // https://stackoverflow.com/questions/664014/what-integer-hash-function-are-good-that-accepts-an-integer-hash-key
-    // Constants provide very good distribution - each input bit affects each output bit with ~50% probability.
-    size_t seed = blockKey.uniqueTokens.size() ^ parentHash * UINT64_C(0xbf58476d1ce4e5b9);
-
-    if (parentHash == 0 && blockKey.cacheSaltID)
-    {
-        // Only hashing the cache salt ID for the first block in the sequence
-        seed = __hash64(blockKey.cacheSaltID.value(), seed);
-    }
-
-    for (auto const& uniqueToken : blockKey.uniqueTokens)
-    {
-        seed = __hash32(static_cast<uint32_t>(uniqueToken.tokenId), seed);
-        if (blockKey.usesExtraIds)
-        {
-            seed = __hash64(uniqueToken.tokenExtraId, seed);
-        }
-    }
-
-    if (blockKey.loraTaskId)
-    {
-        seed = __hash64(blockKey.loraTaskId.value(), seed);
-    }
-
-    // Add extra keys for multimodal data mixing in external multimodal item hash and token offset within this sequence
-    // block
-    if (!blockKey.extraKeys.empty())
-    {
-        for (auto const& mmKey : blockKey.extraKeys)
-        {
-            auto const& mmHash = mmKey.hash;
-            auto const& startOffset = mmKey.startOffset;
-            // Hash the multimodal hash array in 32-bit chunks (more efficient)
-            for (size_t i = 0; i < 32; i += 4)
-            {
-                // Combine 4 bytes into a 32-bit word (construct as little endian order)
-                uint32_t word = static_cast<uint32_t>(mmHash[i]) | (static_cast<uint32_t>(mmHash[i + 1]) << 8)
-                    | (static_cast<uint32_t>(mmHash[i + 2]) << 16) | (static_cast<uint32_t>(mmHash[i + 3]) << 24);
-
-                // Mix the word into the seed
-                seed = __hash32(word, seed);
-            }
-
-            // Hash the start offset
-            seed = __hash64(static_cast<uint64_t>(startOffset), seed);
-        }
-    }
-
-    return seed;
-}
-
 KVCacheBlock::KVCacheBlock(IdType blockId, tk::KVCacheIndex blockIdx)
     : mBlockId(blockId)
     , mMemoryPoolBlockIndex{blockIdx}
